Move kubeclient to clusterman
