+++ cloudman/clusterman/clients/helpers.py
